@@ -26,7 +26,8 @@
         private readonly IEqualityComparer<string>? _comparer;
         private readonly FileInfo _fileInfo;
         private readonly string _filePath;
-        private readonly JsonSerializerOptions _jsonSerializerOptions;
+        private readonly JsonSerializerOptions _serializerOptions;
+        private readonly JsonSerializerOptions _deserializerOptions;
         private Dictionary<string, Dictionary<string, TEntity>> _entities;
         private DateTime _lastModified = DateTime.MinValue;
 
@@ -43,18 +44,31 @@
             _comparer = comparer;
             _entities = new Dictionary<string, Dictionary<string, TEntity>>(_comparer);
             _fileInfo = new FileInfo(_filePath);
-            _jsonSerializerOptions = new JsonSerializerOptions
+
+            // Serialization options
+            _serializerOptions = new JsonSerializerOptions
             {
                 DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull,
                 WriteIndented = true
             };
-            _jsonSerializerOptions.Converters.Add(new JsonStringEnumConverter());
-            _jsonSerializerOptions.Converters.Add(new ObjectJsonConverter());
+            _serializerOptions.Converters.Add(new JsonStringEnumConverter());
             if (converters != null)
             {
                 foreach (var converter in converters)
                 {
-                    _jsonSerializerOptions.Converters.Add(converter);
+                    _serializerOptions.Converters.Add(converter);
+                }
+            }
+
+            // Deserialization options
+            _deserializerOptions = new JsonSerializerOptions();
+            _deserializerOptions.Converters.Add(new JsonStringEnumConverter());
+            _deserializerOptions.Converters.Add(new ObjectJsonConverter());
+            if (converters != null)
+            {
+                foreach (var converter in converters)
+                {
+                    _deserializerOptions.Converters.Add(converter);
                 }
             }
         }
@@ -404,7 +418,7 @@
         private void Serialize()
         {
             using var streamWriter = new StreamWriter(_filePath);
-            var json = JsonSerializer.Serialize(_entities, _jsonSerializerOptions);
+            var json = JsonSerializer.Serialize(_entities, _serializerOptions);
             streamWriter.Write(json);
         }
 
@@ -417,7 +431,7 @@
 
             using var streamReader = new StreamReader(_filePath);
             var json = streamReader.ReadToEnd();
-            var entities = JsonSerializer.Deserialize<Dictionary<string, Dictionary<string, TEntity>>>(json, _jsonSerializerOptions);
+            var entities = JsonSerializer.Deserialize<Dictionary<string, Dictionary<string, TEntity>>>(json, _deserializerOptions);
             _entities = new Dictionary<string, Dictionary<string, TEntity>>(_comparer);
             foreach (var group in entities!)
             {
